requirements/prod.txt
Django==5.2.12
PyJWT==2.12.1
celery[redis]==5.6.3
django-debug-toolbar==6.3.0
django-redis==5.2.0
flower==2.0.1
psycopg2-binary==2.9.12
python-dotenv==1.2.2
requests==2.33.0
whitenoise==6.12.0
yandexcloud==0.383.0